+++ Source/LibationWinForms/ProcessQueue/SidebarControl.resx
